fix: cache race condition after Service creation
We are no longer sharing a cache between server and controllers.
This unfortunately allowed for a race condition after Service creation.
Controller might be missing Service resource in vcluster cache,
which is causing the deletion of the Service in the host cluster,
until it is recreated again by the controller based on Service resource
that gets eventually pulled into vcluster cache.

To prevent this race condition we leave the annotation
that prevents Service deletion in the host until the Service resource
in vcluster is in the expected state.