Retain the PersistentVolume when the PersistentVolumeReclaimPolicy is 'Retain'
A check for the PersistentVolumeReclaimPolicy of the PersistentVolume has been introduced
in the syncer as part of the fix. If the PersistentVolumeReclaimPolicy is Retain,
then the PersistentVolume is not deleted. Now, when the persistent volume is retained,
the names of the Claim and StorageClass was displayed from the Physical PersistentVolume,
the translateUpdateBackwards function in the translator has been modified to get the names
from the Virtual PersistentVolume, in case the PersistentVolumeClaim no longer exists.