fix(chart): ignoring .securityContext values for coredns
The values were hardcoded in the go code, but exposed as a go template.
Hardcoded the value in the helm chart to clarify that this is not
configurable for coredns (root is not needed).

The EKS chart was not rendering the go template properly as it was not
escaped properly.

Signed-off-by: Rohan CJ <rohantmp@gmail.com>