Fix: vclusterctl - Wrong connect logic for the list command
The `list` command don't correctly return if virtual cluster is connected when
applying the global context filtering (option `--context`) in the `table`
output format.

The `json` output format was also missing the `connected` and `age` fields that
are available in the `table` format.

The solution for the connected logic is always run the routine to detect the current context, and
use that as reference when looping over the existing virtual clusters. This is
required because `cmd.Context` could contain the filtering context instead of
the connected one.

For the `json` output, the type `VCluster`, that was created on the same file,
was used to build the correct output format. Missing fields were added to it.