Update stopwatch.go
Hi this is my submission for hacktoberfest 2023!
Minimize Logger Calls: You can minimize the number of logger calls by using the logger only when necessary. In some cases, you may not want to call the logger, especially if the elapsed time is very short.

Use a Timer Interface: Create an interface for timing functionality to make it easier to test and add new features in the future.

Add a Stop Timer Function: Implement a Stop function that accepts an optional callback function that can be executed upon stopping the timer. This allows you to perform custom actions when stopping the timer.

Make It Go-routine Safe: If you plan to use the timer in concurrent go routines, you should ensure that the timer is go routine-safe.

This code maintains the functionality of measuring time and logging but allows you to customize the logging and execution of callback functions based on elapsed time.