Connect to vCluster Platform earlier
vCluster is now able to request a data source from vCluster Platform.
vCluster is only able to use the data source if it attempts to connect
to the vCluster Platform prior to deploying its own control plane.
vCluster Platform is now connected before control plane is deployed
but the platform controllers and tailscale server are started after
because they require the control plane to be running.