Check if vcluster version supports k8s version
Prior, an extra value, distro.k8s.version, could be set for a vcluster
with a version that does not support the distro.k8s.version field.
This would lead to a panic in vcluster. Now, the value is only set if
the vcluster chart version is known to support it. Otherwise, it is
left empty, leading to the omission of the field and no panic.

fix