Only uninstall if reset flag is true
Prior, the reset flag indicated that resources should be uninstalled.
However, if an existing deployment was not detected, all other
resources were removed regardless of the reset flag's value. Now,
reset must be set to true to uninstall any existing vCluster Platform
resources. This is more in line with how a reinstall performed with
helm behaves and is more predictable for the user. The usage
description has been updated to reflect the change.