Remove use of client from CLI setup
Prior, we were using a kubeconfig client during setup of the platform
commands. The client uses the command "vcluster platform token". This
would cause an endless loop like the following:
1. you run the "vcluster platform ..." command
2. vcluster makes request to /deployments endpoint, before running
the intended command.
3. the k8s client uses vcluster to authenticate the request
4. return to step 2
The request was being made to detect the platform namespace and set
it as the default namespace for the backup management command. Now,
there is not a default that is set ahead of running the command but
if a value was not passed, the CLI will then search for the platform
namespace.

(cherry picked from commit a5622f963c27200bc8ceab904d7a6079c62904ad)