Delete deployment label selector is outdated (#2369)
Prior, if the coredns deployment was using the old label, "kube-dns",
it would not be updated during the coredns manifests apply. This was
because label selectors are immutable. Now, if the label is outdated
on the deployment, the deployment is deleted prior to the coredns
manifests apply.

(cherry picked from commit 595cf00dabde578acd07024ef476346b702e176f)