fix: revert networkpolicy label selector
Prior, the label selector "k8s-app=kube-dns" in the NetworkPolicy
template was changed to target the updated vcluster coredns pod name,
vcluster-kube-dns. Enabling NetworkPolicy creates a NetworkPolicy in
the host cluster. The NetworkPolicy intends to target the host's
coredns pod which is still named kube-dns. The NetworkPolicy is
additionally targeting the kube-system namespace. There is no pod that
matches the "k8s-app=kube-dns" selector in the host's kube-system
namespace. Now, the change to the NetworkPolicy's label selector has
been reverted so that it properly selects for the host's kube-dns
pod.