Allow embedded to external backing store switch
This exception is made due to a bug that caused backing stores to
be annotated as embedded but use external data sources and connectors.
This happened when these external fields were provided without
setting the external enabled flag. A change was made where not
setting the enabled flag would correctly cause vCluster to default
to using an embedded (SQLite) data store. After upgrading this could
cause a vCluster to switch from using an external data store to an
embedded data store. Now, users can update the external data store
enabled flag to be true even though it was recognized as an embedded
cluster before.